--- app/src/main/java/com/atreus/asisapp/Account/ForgotPasswordActivity.kt
@@ -1,49 +1,0 @@
-package com.atreus.asisapp.Account
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.util.Patterns
-import android.widget.Toast
-import com.atreus.asisapp.databinding.ActivityForgotPasswordBinding
-import com.google.firebase.auth.FirebaseAuth
-
-class ForgotPasswordActivity : AppCompatActivity() {
-    lateinit var binding : ActivityForgotPasswordBinding
-    lateinit var auth : FirebaseAuth
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        binding = ActivityForgotPasswordBinding.inflate(layoutInflater)
-        super.onCreate(savedInstanceState)
-        setContentView(binding.root)
-
-        binding.btnReset.setOnClickListener{
-            val email = binding.edtEmailReset.text.toString()
-            val edtEmail = binding.edtEmailReset
-
-            if (email.isEmpty()) {
-                edtEmail.error = "Email tidak boleh kosong"
-                edtEmail.requestFocus()
-                return@setOnClickListener
-            }
-
-            if (!Patterns.EMAIL_ADDRESS.matcher(email).matches()){
-                edtEmail.error = "Email tidak valid"
-                edtEmail.requestFocus()
-                return@setOnClickListener
-            }
-
-            FirebaseAuth.getInstance().sendPasswordResetEmail(email).addOnCompleteListener{
-                if(it.isSuccessful){
-                    Toast.makeText(this, "Email reset password telah terkirim", Toast.LENGTH_SHORT).show()
-                    val intent = Intent(this, LoginActivity::class.java)
-                    startActivity(intent)
-                    finish()
-                } else {
-                    Toast.makeText(this, "${it.exception?.message}", Toast.LENGTH_SHORT).show()
-                }
-            }
-
-        }
-    }
-}
