--- app/src/main/java/com/atreus/asisapp/Home/Fragment/Habit.kt
@@ -1,22 +1,0 @@
-package com.atreus.asisapp.Home.Fragment
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.atreus.asisapp.R
-
-
-class Habit : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_habit, container, false)
-    }
-
-
-}
