--- app/src/main/java/com/atreus/asisapp/Home/HomeActivity.kt
@@ -1,48 +1,0 @@
-package com.atreus.asisapp.Home
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import com.atreus.asisapp.Home.Fragment.Habit
-import com.atreus.asisapp.Home.Fragment.Home
-import com.atreus.asisapp.Home.Fragment.Profile
-import com.atreus.asisapp.Home.Fragment.Work
-import com.atreus.asisapp.R
-import com.atreus.asisapp.databinding.ActivityHomeBinding
-
-class HomeActivity : AppCompatActivity() {
-    private lateinit var binding : ActivityHomeBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityHomeBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-        replaceFragment(Home())
-
-        binding.bottomNavigationView.setOnItemSelectedListener {
-
-            when(it.itemId){
-
-                R.id.nav_home -> replaceFragment(Home())
-                R.id.nav_work -> replaceFragment(Work())
-                R.id.nav_habit -> replaceFragment(Habit())
-                R.id.nav_profile -> replaceFragment(Profile())
-
-                else ->{
-
-                    }
-
-                }
-                true
-            }
-
-        }
-
-    private fun replaceFragment(fragment: Fragment){
-
-        val fragmentManager = supportFragmentManager
-        val fragmentTransaction = fragmentManager.beginTransaction()
-        fragmentTransaction.replace(R.id.frame_layout,fragment)
-        fragmentTransaction.commit()
-    }
-}
