+++ app/src/main/java/com/atreus/asisapp/account/LoginActivity.kt
@@ -1,0 +1,75 @@
+package com.atreus.asisapp.account
+
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.util.Patterns
+import android.widget.Toast
+import com.atreus.asisapp.home.HomeActivity
+import com.atreus.asisapp.databinding.ActivityLoginBinding
+import com.google.firebase.auth.FirebaseAuth
+
+
+class LoginActivity : AppCompatActivity() {
+    lateinit var binding: ActivityLoginBinding
+    lateinit var auth: FirebaseAuth
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        binding = ActivityLoginBinding.inflate(layoutInflater)
+        super.onCreate(savedInstanceState)
+        setContentView(binding.root)
+
+        auth = FirebaseAuth.getInstance()
+
+        binding.forgotPassword.setOnClickListener{
+            val intent = Intent(this, ForgotPasswordActivity::class.java)
+            startActivity(intent)
+        }
+
+        binding.daftarBaru.setOnClickListener {
+            val intent = Intent(this, RegisterActivity::class.java)
+            startActivity(intent)
+        }
+
+        binding.buttonLogin.setOnClickListener {
+            val email = binding.usernameLogin.text.toString()
+            val password = binding.passwordLogin.text.toString()
+
+            //Validasi email
+            if (email.isEmpty()) {
+                binding.usernameLogin.error = "Email Harus Diisi"
+                binding.usernameLogin.requestFocus()
+                return@setOnClickListener
+            }
+
+            //Validasi email tidak sesuai
+            if (!Patterns.EMAIL_ADDRESS.matcher(email).matches()) {
+                binding.usernameLogin.error = "Email tidak valid"
+                binding.usernameLogin.requestFocus()
+                return@setOnClickListener
+            }
+
+            //Validasi password
+            if (password.isEmpty()) {
+                binding.passwordLogin.error = "Password Harus Diisi"
+                binding.passwordLogin.requestFocus()
+                return@setOnClickListener
+            }
+
+            RegisterFirebase(email, password)
+        }
+    }
+
+    private fun RegisterFirebase(email: String, password: String) {
+        auth.signInWithEmailAndPassword(email, password)
+            .addOnCompleteListener(this) {
+                if (it.isSuccessful) {
+                    Toast.makeText(this, "Selamat datang", Toast.LENGTH_SHORT).show()
+                    val intent = Intent(this, HomeActivity::class.java)
+                    startActivity(intent)
+                } else {
+                    Toast.makeText(this, "${it.exception?.message}", Toast.LENGTH_SHORT).show()
+                }
+            }
+    }
+}
