+++ app/src/main/java/com/atreus/asisapp/account/RegisterActivity.kt
@@ -1,0 +1,70 @@
+package com.atreus.asisapp.account
+
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.util.Patterns
+import android.widget.Toast
+import com.atreus.asisapp.databinding.ActivityRegisterBinding
+import com.google.firebase.auth.FirebaseAuth
+
+class RegisterActivity : AppCompatActivity() {
+    lateinit var binding : ActivityRegisterBinding
+    lateinit var auth : FirebaseAuth
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        binding = ActivityRegisterBinding.inflate(layoutInflater)
+        super.onCreate(savedInstanceState)
+        setContentView(binding.root)
+
+        auth = FirebaseAuth.getInstance()
+
+        binding.btnBuatAkun.setOnClickListener{
+            val email = binding.etEmail.text.toString()
+            val password = binding.etPassword.text.toString()
+
+            //Validasi email
+            if (email.isEmpty()){
+                binding.etEmail.error = "Email Harus Diisi"
+                binding.etEmail.requestFocus()
+                return@setOnClickListener
+            }
+
+            //Validasi email tidak sesuai
+            if (!Patterns.EMAIL_ADDRESS.matcher(email).matches()){
+                binding.etEmail.error = "Email tidak valid"
+                binding.etEmail.requestFocus()
+                return@setOnClickListener
+            }
+
+            //Validasi password
+            if (password.isEmpty()){
+                binding.etPassword.error = "Password Harus Diisi"
+                binding.etPassword.requestFocus()
+                return@setOnClickListener
+            }
+
+            //Validasi email tidak sesuai
+            if (password.length < 8){
+                binding.etEmail.error = "Minimal password 6 karakter"
+                binding.etEmail.requestFocus()
+                return@setOnClickListener
+            }
+
+            RegisterFirebase(email, password)
+        }
+    }
+
+    private fun RegisterFirebase(email: String, password: String) {
+        auth.createUserWithEmailAndPassword(email, password)
+            .addOnCompleteListener(this){
+                if (it.isSuccessful){
+                    Toast.makeText(this, "Register berhasil", Toast.LENGTH_SHORT ).show()
+                    val intent = Intent(this, LoginActivity::class.java)
+                    startActivity(intent)
+                } else {
+                    Toast.makeText(this, "${it.exception?.message}", Toast.LENGTH_SHORT).show()
+                }
+            }
+    }
+}
