+++ app/src/main/java/com/atreus/asisapp/onboarding/ViewPagerAdapter.kt
@@ -1,0 +1,24 @@
+package com.atreus.asisapp.onboarding
+
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.lifecycle.Lifecycle
+import androidx.viewpager2.adapter.FragmentStateAdapter
+
+class ViewPagerAdapter (
+    list : ArrayList<Fragment>,
+    fm : FragmentManager,
+    lifecycle: Lifecycle
+): FragmentStateAdapter(fm , lifecycle) {
+
+    private val fragmentList:ArrayList<Fragment> = list
+
+    override fun getItemCount(): Int {
+        return fragmentList.size
+    }
+
+    override fun createFragment(position: Int): Fragment {
+        return fragmentList[position]
+    }
+
+}
