+++ app/src/main/res/drawable/circular_progress_bar.xml
@@ -1,0 +1,52 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--<rotate-->
+<!--    xmlns:android="http://schemas.android.com/apk/res/android"-->
+<!--    android:fromDegrees="270"-->
+<!--    android:toDegrees="270">-->
+
+    <!--styling the progress bar-->
+    <layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+        <item>
+            <rotate android:fromDegrees="270" android:toDegrees="270">
+                <shape
+                    android:innerRadiusRatio="2.5"
+                    android:shape="ring"
+                    android:thickness="20dp"
+                    android:useLevel="true">
+                    <gradient
+                        android:angle="20"
+                        android:endColor="@color/white"
+                        android:startColor="@color/white"
+                        android:type="sweep"
+                        android:useLevel="false" />
+                </shape>
+
+            </rotate>
+        </item>
+        <item android:bottom="185dp">
+                <shape
+                    android:innerRadiusRatio="1000"
+                    android:shape="ring"
+                    android:thickness="10dp"
+                    android:useLevel="false">
+                    <solid android:color="@color/white" />
+                </shape>
+        </item>
+        <item>
+            <rotate>
+                <inset android:insetBottom="185dp">
+                    <shape
+                        android:innerRadiusRatio="1000"
+                        android:shape="ring"
+                        android:thickness="10dp"
+                        android:useLevel="false">
+                        <solid android:color="@color/white" />
+                    </shape>
+                </inset>
+            </rotate>
+        </item>
+
+    </layer-list>
+
+
+<!--</rotate>-->
