+++ app/src/main/res/drawable/circular_shape.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:innerRadiusRatio="2.5"
+    android:shape="ring"
+    android:thickness="20dp"
+    android:useLevel="false">
+    <solid android:color="#33FFFFFF" />
+</shape>
