+++ app/src/main/res/drawable/outline_blue.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:aapt="http://schemas.android.com/aapt"
+    android:width="116dp"
+    android:height="56dp"
+    android:viewportWidth="116"
+    android:viewportHeight="56"
+    >
+
+    <group>
+
+        <clip-path
+            android:pathData="M8 0H108C112.418 0 116 3.58172 116 8V48C116 52.4183 112.418 56 108 56H8C3.58172 56 0 52.4183 0 48V8C0 3.58172 3.58172 0 8 0Z"
+            />
+
+        <path
+            android:pathData="M0 0V56H116V0"
+            android:fillColor="#FFFFFF"
+            />
+
+        <path
+            android:pathData="M8 0H108C112.418 0 116 3.58172 116 8V48C116 52.4183 112.418 56 108 56H8C3.58172 56 0 52.4183 0 48V8C0 3.58172 3.58172 0 8 0Z"
+            android:strokeWidth="5"
+            android:strokeColor="#1D35B5"
+            />
+
+    </group>
+
+</vector>
