+++ app/src/main/res/drawable/pop_up_pomo.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:aapt="http://schemas.android.com/aapt"
+    android:width="328dp"
+    android:height="269dp"
+    android:viewportWidth="328"
+    android:viewportHeight="269"
+    >
+
+    <group>
+
+        <clip-path
+            android:pathData="M20 0H308C319.046 0 328 8.95431 328 20V249C328 260.046 319.046 269 308 269H20C8.95431 269 0 260.046 0 249V20C0 8.95431 8.95431 0 20 0Z"
+            />
+
+        <path
+            android:pathData="M0 0V269H328V0"
+            android:fillColor="#FFFFFF"
+            />
+
+    </group>
+
+</vector>
