+++ app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/yellow"
-    tools:context=".Account.LoginActivity">
+    tools:context=".account.LoginActivity">
 
     <androidx.cardview.widget.CardView
         android:id="@+id/card_view"
