+++ app/src/main/res/layout/activity_register.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/blue"
     android:backgroundTint="@color/blue"
-    tools:context=".Account.RegisterActivity">
+    tools:context=".account.RegisterActivity">
 
     <TextView
         android:id="@+id/textView"
