+++ app/src/main/res/layout/fragment_first_screen.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".OnBoarding.Screen.FirstScreen"
+    tools:context=".onboarding.Screen.FirstScreen"
     android:background="@color/yellow">
