+++ app/src/main/res/layout/fragment_fourth_screen.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".OnBoarding.Screen.FourthScreen"
+    tools:context=".onboarding.Screen.FourthScreen"
     android:background="@color/cyan">
