+++ app/src/main/res/layout/fragment_habit.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Home.Fragment.Habit">
+    tools:context=".home.Fragment.Habit">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
