+++ app/src/main/res/layout/fragment_profile.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/white"
     android:padding="16dp"
-    tools:context=".Home.Fragment.Profile">
+    tools:context=".home.Fragment.Profile">
 
     <androidx.cardview.widget.CardView
         android:id="@+id/cardView"
