+++ app/src/main/res/layout/fragment_second_screen.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".OnBoarding.Screen.SecondScreen"
+    tools:context=".onboarding.Screen.SecondScreen"
     android:background="@color/blue">
