+++ app/src/main/res/layout/fragment_thrid_screen.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".OnBoarding.Screen.ThridScreen"
+    tools:context=".onboarding.Screen.ThridScreen"
     android:background="@color/pink">
