+++ app/src/main/res/layout/fragment_view_pager.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".OnBoarding.ViewPagerFragment">
+    tools:context=".onboarding.ViewPagerFragment">
 
 
     <androidx.viewpager2.widget.ViewPager2
