+++ app/src/main/res/layout/fragment_work.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Home.Fragment.Work">
+    tools:context=".home.Fragment.Work">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
