+++ app/src/main/res/layout/layout_dialog.xml
@@ -1,0 +1,102 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="@drawable/pop_up_pomo"
+    android:padding="16dp"
+    app:circularflow_radiusInDP="0">
+
+    <TextView
+        android:id="@+id/textView3"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="108dp"
+        android:layout_marginTop="20dp"
+        android:layout_marginEnd="108dp"
+        android:fontFamily="@font/poppins_bold"
+        android:text="Customize Timer"
+        android:textColor="#1D35B5"
+        android:textSize="14sp"
+        android:textStyle="bold"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <EditText
+        android:id="@+id/edit_waktu"
+        android:layout_width="80dp"
+        android:layout_height="48dp"
+        android:layout_marginStart="40dp"
+        android:layout_marginTop="20dp"
+        android:layout_marginEnd="40dp"
+        android:background="@drawable/outline_blue"
+        android:fontFamily="@font/poppins"
+        android:hint="08  :  00"
+        android:inputType="numberDecimal"
+        android:minHeight="24dp"
+        android:textAlignment="center"
+        android:textColor="#1D35B4"
+        android:textColorHint="#1D35B4"
+        android:textSize="12sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toEndOf="@+id/textView2"
+        app:layout_constraintTop_toBottomOf="@+id/textView3" />
+
+    <EditText
+        android:id="@+id/edit_break"
+        android:layout_width="80dp"
+        android:layout_height="48dp"
+        android:background="@drawable/outline_blue"
+        android:fontFamily="@font/poppins"
+        android:hint="08  :  00"
+        android:inputType="numberDecimal"
+        android:minHeight="24dp"
+        android:textAlignment="center"
+        android:textColor="#1D35B4"
+        android:textColorHint="#1D35B4"
+        android:textSize="12sp"
+        app:layout_constraintEnd_toEndOf="@+id/edit_waktu"
+        app:layout_constraintStart_toStartOf="@+id/edit_waktu"
+        app:layout_constraintTop_toTopOf="@+id/textView5" />
+
+    <TextView
+        android:id="@+id/textView2"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="40dp"
+        android:layout_marginTop="15dp"
+        android:fontFamily="@font/poppins_bold"
+        android:text="Pomodoro Timer"
+        android:textColor="#1D35B5"
+        android:textSize="12sp"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="@+id/edit_waktu" />
+
+    <TextView
+        android:id="@+id/textView5"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="40dp"
+        android:layout_marginTop="60dp"
+        android:fontFamily="@font/poppins_bold"
+        android:text="Break Time"
+        android:textColor="#1D35B5"
+        android:textSize="12sp"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/textView2" />
+
+    <Button
+        android:id="@+id/save_btn"
+        android:layout_width="144dp"
+        android:layout_height="60dp"
+        android:layout_marginTop="200dp"
+        android:fontFamily="@font/poppins_bold"
+        android:text="Save !"
+        android:textSize="19sp"
+        app:cornerRadius="10dp"
+        app:layout_constraintEnd_toEndOf="@+id/textView3"
+        app:layout_constraintStart_toStartOf="@+id/textView3"
+        app:layout_constraintTop_toBottomOf="@+id/textView3" />
+</androidx.constraintlayout.widget.ConstraintLayout>
